ACTIN-1326: Move under example package
