+++ system/src/test/kotlin/com/hartwig/actin/system/example/ExampleFunctions.kt
@@ -1,4 +1,4 @@
-package com.hartwig.actin.system
+package com.hartwig.actin.system.example
 
 import com.hartwig.actin.configuration.AlgoConfiguration
 import com.hartwig.actin.configuration.EnvironmentConfiguration
